Reading and writing changes to update xml file after each journey.
Instead of getting each journey from the API and writing to xml file
at the end of the operation, write each journey to xml as it is
fetched.
Also, it can now pick up from where it left off by reading the xml
file at start.
This means that losing the connection will not require a complete
restart but can begin again from where it was previously.